Move Logger to Core project
